+++ .github/CHANGELOG.md
@@ -127,6 +127,7 @@
 1. [A32NX/FWS] Fix master warning / caution lights sometimes not extinguishing properly - @FozzieHi (fozzie)
 1. [FMS] Fixed an issue where waypoints around lateral discontinuities and at enroute/procedure boundaries were sometimes not sent to the ND - @tracernz (Mike)
 1. [FMS] Fixed an issue where the WPT overlay on the ND showed terminal waypoints in enroute range - @tracernz (Mike)
+1. [A32NX/FMS] Limit step descent guidance to -1000fpm - @Jonny23787 (Jonathan)
 
 ## 0.14.0